implement "Add products" functionality:
I have added functiionality of adding a product to the list of products.
WHen you fill the form and submit it, the products gets added to the
list. The product is to json onject and it rereders to reflect on the
UI. Similarly, when delete button is clicked. it removes it from json
and also from UI.